feat: implemented dynamic 404 for all missing pages
